Catch the row exceptions from the preview presenter
Refs #34240
