+++ qt/scientific_interfaces/ISISReflectometry/Reduction/Batch.cpp
@@ -5,6 +5,7 @@
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
 #include "Batch.h"
+#include "Reduction/RowExceptions.h"
 
 #include <utility>
