+++ qt/scientific_interfaces/ISISReflectometry/Reduction/Experiment.h
@@ -60,7 +60,7 @@ public:
   boost::optional<LookupRow> findLookupRow(PreviewRow const &previewRow, double tolerance) const;
   boost::optional<LookupRow> findWildcardLookupRow() const;
 
-  void replaceLookupRow(PreviewRow const &previewRow, double tolerance);
+  void replaceLookupRow(LookupRow lookupRow, double tolerance);
 
   boost::optional<size_t> getLookupRowIndexFromRow(Row const &row, double tolerance) const;
