+++ qt/scientific_interfaces/ISISReflectometry/Reduction/LookupTable.h
@@ -8,26 +8,14 @@
 #pragma once
 
 #include "Common/DllConfig.h"
-#include "GUI/Preview/ROIType.h"
 #include "LookupRow.h"
 #include <boost/optional.hpp>
 #include <boost/regex.hpp>
-#include <exception>
 #include <vector>
 
 namespace MantidQt::CustomInterfaces::ISISReflectometry {
 class Row;
 class PreviewRow;
-
-struct RowNotFoundException : public std::out_of_range {
-public:
-  RowNotFoundException(std::string s) : std::out_of_range(std::move(s)){};
-};
-
-struct MultipleRowsFoundException : public std::length_error {
-public:
-  MultipleRowsFoundException(std::string s) : std::length_error(std::move(s)){};
-};
 
 class MANTIDQT_ISISREFLECTOMETRY_DLL LookupTable {
 public:
@@ -38,7 +26,7 @@ public:
   boost::optional<LookupRow> findLookupRow(Row const &row, double tolerance) const;
   boost::optional<LookupRow> findLookupRow(PreviewRow const &previewRow, double tolerance) const;
   boost::optional<LookupRow> findWildcardLookupRow() const;
-  void replaceLookupRow(PreviewRow const &previewRow, double tolerance);
+  void replaceLookupRow(LookupRow lookupRow, double tolerance);
   size_t getIndex(LookupRow const &) const;
   std::vector<LookupRow> const &rows() const;
   std::vector<LookupRow::ValueArray> toValueArray() const;
@@ -54,8 +42,6 @@ private:
   std::vector<LookupRow> searchByTitle(Row const &row) const;
   std::vector<LookupRow> findMatchingRegexes(std::string const &title) const;
   std::vector<LookupRow> findEmptyRegexes() const;
-
-  void replaceLookupRowProcessingInstructions(PreviewRow const &previewRow, LookupRow &lookupRow, ROIType regionType);
 };
 
 } // namespace MantidQt::CustomInterfaces::ISISReflectometry
